+++ example/lib/views/app_text_field.dart
@@ -7,6 +7,7 @@ class AppTextField extends StatelessWidget {
   final bool isNextButton;
   final bool isPrevious;
   final bool showDoneKeyboard;
+  final ValueNotifier<bool>? showExtraHeight;
   const AppTextField({
     super.key,
     required this.focusNode,
@@ -14,34 +15,38 @@ class AppTextField extends StatelessWidget {
     this.isNextButton = false,
     this.isPrevious = false,
     this.showDoneKeyboard = false,
+    this.showExtraHeight,
   });
 
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.all(20.0),
-      child: TextFormField(
-        focusNode: showDoneKeyboard
-            ? (focusNode
-              ..addListener(() {
-                final hasFocus = focusNode.hasFocus;
-                if (hasFocus) {
-                  KeyboardOverlay.showOverlay(
-                    context: context,
-                    isNextButton: isNextButton,
-                    isPrevious: isPrevious,
-                    isShowButton: (isNextButton || isPrevious),
-                  );
-                } else {
-                  KeyboardOverlay.removeOverlay();
-                }
-              }))
-            : focusNode,
-        keyboardType: TextInputType.text,
-        decoration: InputDecoration(
-          labelText: label,
-        ),
-      ),
-    );
+        padding: const EdgeInsets.all(20.0),
+        child: ValueListenableBuilder(
+            valueListenable: showExtraHeight!,
+            builder: (context, value, child) => TextFormField(
+                  focusNode: showDoneKeyboard
+                      ? (focusNode
+                        ..addListener(() {
+                          final hasFocus = focusNode.hasFocus;
+                          if (hasFocus) {
+                            showExtraHeight?.value = true;
+                            KeyboardOverlay.showOverlay(
+                              context: context,
+                              isNextButton: isNextButton,
+                              isPrevious: isPrevious,
+                              isShowButton: (isNextButton || isPrevious),
+                            );
+                          } else {
+                            KeyboardOverlay.removeOverlay();
+                            showExtraHeight?.value = false;
+                          }
+                        }))
+                      : focusNode,
+                  keyboardType: TextInputType.text,
+                  decoration: InputDecoration(
+                    labelText: label,
+                  ),
+                )));
   }
 }